Moved layout code to own module
