+++ src/layout/layout_tree/mod.rs
@@ -1,3 +1,6 @@
+pub mod layout;
 pub mod container;
 pub mod tree;
 mod graph_tree;
+
+use self::tree::LayoutTree;
